Added a despawner of the platforms, added a GameOverScene, and added scenes to the build.
For the despawner I attached a script to the platform prefab that will keep track of the bottom pointer and check the y position of each platform it encounters and will destroy it when it reaches it. For the game over scene I attached a script that keeps track of the y position of the player and if it becomes less than the bottom pointer, if it does then the player died and moves on to the game over scene.